show number of have voted delegates
