+++ app/Model/Delegate/ReadModel/Queries/VotedDelegatesCountQuery.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Model\Delegate\ReadModel\Queries;
+
+use Model\Delegate\ReadModel\QueryHandlers\VotedDelegatesCountQueryHandler;
+
+/**
+ * @see VotedDelegatesCountQueryHandler
+ */
+class VotedDelegatesCountQuery
+{
+}
